backend/requirements.txt
fastapi
google.generativeai
dotenv
python-multipart
pillow
PyMuPDF
fastapi[standard]